backend/requirements.txt
# Core dependencies
pandas>=1.5.0,<2.2.0
numpy>=1.21.0,<2.0.0
scipy>=1.9.0,<2.0.0
matplotlib>=3.5.0,<4.0.0

# Web scraping
selenium>=4.0.0,<5.0.0
requests>=2.28.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
beautifulsoup4>=4.11.0,<5.0.0
feedparser>=6.0.0,<7.0.0

# Database
mysql-connector-python>=8.0.0,<9.0.0

# AI and Agentic Framework
langgraph>=0.1.0,<1.0.0
openai>=1.58.1,<2.0.0
crewai>=0.140.0,<1.0.0
crewai-tools>=0.49.0,<1.0.0
langchain-openai>=0.2.14,<1.0.0

# Text processing and analysis
textblob>=0.17.0,<1.0.0

# Web framework
flask>=2.3.0,<3.0.0
flask-cors>=4.0.0,<5.0.0

# Logging and utilities
python-dateutil>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytz>=2023.3,<2024.0

# Additional utilities
aiofiles>=23.0.0,<24.0.0

# AI Assistant
groq>=0.4.0,<1.0.0

# Development dependencies (optional)
# pytest>=7.0.0,<8.0.0
# black>=23.0.0,<24.0.0
# flake8>=6.0.0,<7.0.0